- DoF Client Profile(Character) -> Options all the checkboxes set the wrong thing (afk, roleplaying, camping, linkdead, lfg, so on) - DoF client Quest complete reward now includes temp rewards item data, previously you would see an empty quest reward window. - DoF client properly in combat (no re-sheething of weapon while in combat) - DoF client fixed spells to do start and end cast at appropriate times (struct fix). - DoF client casting on self says "not a friend" -- now says not an enemy.  When being too far away, instead of saying "too close" now says "too far away". - DoF fixed selecting self interrupts /camp - DoF client server/client properly synched for inventory (food/drink and other equipped slots previously not working right) - DoF Broker no longer crashes client, limited implementation to get first 8 results. WIP - DoF fixed examine equipped items didn't work for drink/potentially mismatched other slots - DoF item stat review  of resists now properly display - DoF client is constantly sending HandleExamineInfoRequest packets for spells, fixed. Greatly improved performance of DoF cause it was constantly sending these packets for all spells in your book per second! - DoF collections window is fixed, collections now display correctly in journal window. Upon turning in complete collections, client no longer crashes. Outstanding note: DoF client has EXP show up as % of level (60?) instead of just bare XP points. This might need more discovery, but isn't a huge issue right now. - DoF player profile inspection inventory, DoF and AoM both display inventory (DoF does not support appearance gear). Cross client inspection works also. Avoidance (DoF and AOM) / ATK (DoF) stats need to be reviewed and matched in the structure, but not a big priority. - DoF client /who list is broken (if more than one player, by myself /who works) -- structure updated for some unknown bytes tail of packet - DoF client gestures now show up visually and not as an /emote- - DoF item display fixed for house items (the items themselves in inventory/broker/so on were previosuly displayed corrupted) - DoF Merchant "sell to merchant" just shows buy window, sell window was displaying buy items, there is already a sell window with the buy window. - DoF merchant mender/repair no longer crashes the client -- currently worked around by sending buy screen then repair screen.. displays both, need to get buy screen disabled, will be its own git issue. - Remove/Add skills via LUA will now properly update/reflect the database. - starting_skills and starting_spells no longer applies on each login, this will stop the restoring of spells/skills on zoning that are removed with LUA - Login sequence into world now initially load spells/skills (and also sync if first time into the world for the character).  This will allow us to avoid an unneccessary blocking behavior with World::SyncCharAbilities (it waited for 5 loops and hold the thread to see if the DB is updating the character).  Now we enforce this behavior with an updated state machine for the login, no longer blocking/holding the thread. alter table characters add column first_world_login tinyint(1) unsigned not null default 0; - R_World, DisplayItemTiers rule made to display item tiers (1) / disable (0) - Bots are no longer impacted by /depop or /repop - fixed crash with Player::CheckQuestFlag where we try to send updates based on quest, dead quest ptr when calling CheckQuestChatUpdate - revive points now support an always_included option, otherwise it only shows the closest distance - Database updates: alter table revive_points add column always_included tinyint(1) unsigned not null default 0; alter table characters add column first_world_login tinyint(1) unsigned not null default 0;
- LoginServer
	* LoginServer.ini now supports (integer) fields expansionsflag, citiesflag, defaultsubscriptionlevel, enabledraces under the "LoginConfig" block
	eg. enabledraces = 65535 allows all races
	    enabledraces = 57343 removes sarnak as playable race (-8192), another -4096 would remove Arasai.  -2048 removes Fae.

	Default values (in hex, you can only use integer in the ini):
		// full support = 0x7CFF
		// 1 << 12 (-4096) = missing echoes of faydwer, disables Fae and Arasai (black portraits) and kelethin as starting city
		// 1 << 13 (-8192) = disables sarnak (black portraits) and gorowyn as starting city
		expansionFlag = 0x7CFF; // 0x4CF5

		/* dword_1ECBA18 operand for race flag packs (sublevel 0,1,2?) -- (sublevel -1) controls starting zones omission 0xEE vs 0xCF (CF misses halas)
		1 = city of qeynos
		2 = city of freeport
		4 = city of kelethin
		8 = city of neriak
		16 = gorowyn
		32 = new halas
		64 = queens colony
		128 = outpost overlord
		*/
		citiesFlag = 0xFF;

		// sub_level 0xFFFFFFFF = blacks out all portraits for class alignments, considered non membership
		// sub_level > 0 = class alignments still required, but portraits are viewable and race selectable
		// sub_level = 2 membership, you can 'create characters on time locked servers' vs standard
		// sub_level = 0 forces popup on close to web browser
		defaultSubscriptionLevel = 0xFFFFFFFF;

		// disable extra races FAE(16) ARASAI (17) SARNAK (18) -- with 4096/8192 flags, no visibility of portraits
		enabledRaces = 0xFFFF; // 0xCFFF

+++ EQ2/source/WorldServer/Player.h
@@ -511,8 +511,8 @@ public:
 	bool CanEquipItem(Item* item, int8 slot);
 	void SetEquippedItemAppearances();
 	vector<EQ2Packet*>	UnequipItem(int16 index, sint32 bag_id, int8 slot, int16 version, int8 appearance_type = 0, bool send_item_updates = true);
-	int8 ConvertSlotToClient(int8 slot, int16 version);
-	int8 ConvertSlotFromClient(int8 slot, int16 version);
+	int16 ConvertSlotToClient(int8 slot, int16 version);
+	int16 ConvertSlotFromClient(int8 slot, int16 version);
 	EQ2Packet* SwapEquippedItems(int8 slot1, int8 slot2, int16 version, int16 equiptype);
 	EQ2Packet*	RemoveInventoryItem(int8 bag_slot, int8 slot);
 	EQ2Packet*	SendInventoryUpdate(int16 version);
